+++ cpp/oneapi/dal/algo/jaccard/BUILD
@@ -9,6 +9,7 @@ dal_module(
     auto = True,
     dal_deps = [
         "@onedal//cpp/oneapi/dal:core",
+        "@onedal//cpp/oneapi/dal/backend/primitives:intersection",
     ]
 )
