+++ cpp/oneapi/dal/algo/jaccard/backend/cpu/vertex_similarity_default_kernel_avx512.hpp
@@ -23,334 +23,28 @@
 #include "oneapi/dal/algo/jaccard/backend/cpu/vertex_similarity_default_kernel.hpp"
 #include "oneapi/dal/algo/jaccard/common.hpp"
 #include "oneapi/dal/algo/jaccard/vertex_similarity_types.hpp"
+#include "oneapi/dal/algo/jaccard/detail/service.hpp"
 #include "oneapi/dal/backend/dispatcher.hpp"
 #include "oneapi/dal/backend/interop/common.hpp"
 #include "oneapi/dal/backend/interop/table_conversion.hpp"
+#include "oneapi/dal/backend/primitives/intersection/intersection.hpp"
 #include "oneapi/dal/detail/policy.hpp"
 #include "oneapi/dal/graph/detail/service_functions_impl.hpp"
 #include "oneapi/dal/table/detail/table_builder.hpp"
+#include "oneapi/dal/table/homogen.hpp"
 
-namespace oneapi::dal::preview {
-namespace jaccard {
-namespace detail {
+namespace oneapi::dal::preview::jaccard::backend {
 
-#if defined(__INTEL_COMPILER)
-ONEDAL_FORCEINLINE std::int32_t _popcnt32_redef(const std::int32_t &x) {
-    return _popcnt32(x);
-}
-#define GRAPH_STACK_ALING(x) __declspec(align(x))
-#else
-ONEDAL_FORCEINLINE std::int32_t _popcnt32_redef(const std::int32_t &x) {
-    std::int32_t count = 0;
-    std::int32_t a = x;
-    while (a != 0) {
-        a = a & (a - 1);
-        count++;
-    }
-    return count;
-}
-#define GRAPH_STACK_ALING(x) \
-    {}
-#endif
-
-ONEDAL_FORCEINLINE std::int64_t intersection(const std::int32_t *neigh_u,
-                                             const std::int32_t *neigh_v,
-                                             std::int32_t n_u,
-                                             std::int32_t n_v) {
-    std::int64_t total = 0;
-    std::int32_t i_u = 0, i_v = 0;
-#if defined(__INTEL_COMPILER)
-    while (i_u < (n_u / 16) * 16 && i_v < (n_v / 16) * 16) { // not in last n%16 elements
-        // assumes neighbor list is ordered
-        std::int32_t min_neigh_u = neigh_u[i_u];
-        std::int32_t max_neigh_v = neigh_v[i_v + 15];
-
-        if (min_neigh_u > max_neigh_v) {
-            if (min_neigh_u > neigh_v[n_v - 1]) {
-                return total;
-            }
-            i_v += 16;
-            continue;
-        }
-
-        std::int32_t min_neigh_v = neigh_v[i_v];
-        std::int32_t max_neigh_u = neigh_u[i_u + 15];
-        if (min_neigh_v > max_neigh_u) {
-            if (min_neigh_v > neigh_u[n_u - 1]) {
-                return total;
-            }
-            i_u += 16;
-            continue;
-        }
-        __m512i v_u = _mm512_loadu_si512((void *)(neigh_u + i_u)); // load 16 neighbors of u
-        __m512i v_v = _mm512_loadu_si512((void *)(neigh_v + i_v)); // load 16 neighbors of v
-        if (max_neigh_u >= max_neigh_v)
-            i_v += 16;
-        if (max_neigh_u <= max_neigh_v)
-            i_u += 16;
-
-        __mmask16 match = _mm512_cmpeq_epi32_mask(v_u, v_v);
-        if (_mm512_mask2int(match) != 0xffff) { // shortcut case where all neighbors match
-            __m512i circ1 = _mm512_set_epi32(0, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1);
-            __m512i circ2 = _mm512_set_epi32(1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2);
-            __m512i circ3 = _mm512_set_epi32(2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3);
-            __m512i circ4 = _mm512_set_epi32(3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4);
-            __m512i circ5 = _mm512_set_epi32(4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5);
-            __m512i circ6 = _mm512_set_epi32(5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6);
-            __m512i circ7 = _mm512_set_epi32(6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8, 7);
-            __m512i circ8 = _mm512_set_epi32(7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9, 8);
-            __m512i circ9 = _mm512_set_epi32(8, 7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10, 9);
-            __m512i circ10 = _mm512_set_epi32(9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11, 10);
-            __m512i circ11 = _mm512_set_epi32(10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12, 11);
-            __m512i circ12 = _mm512_set_epi32(11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13, 12);
-            __m512i circ13 = _mm512_set_epi32(12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 15, 14, 13);
-            __m512i circ14 = _mm512_set_epi32(13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 15, 14);
-            __m512i circ15 = _mm512_set_epi32(14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0, 15);
-            __m512i v_v1 = _mm512_permutexvar_epi32(circ1, v_v);
-            __m512i v_v2 = _mm512_permutexvar_epi32(circ2, v_v);
-            __m512i v_v3 = _mm512_permutexvar_epi32(circ3, v_v);
-            __m512i v_v4 = _mm512_permutexvar_epi32(circ4, v_v);
-            __m512i v_v5 = _mm512_permutexvar_epi32(circ5, v_v);
-            __m512i v_v6 = _mm512_permutexvar_epi32(circ6, v_v);
-            __m512i v_v7 = _mm512_permutexvar_epi32(circ7, v_v);
-            __m512i v_v8 = _mm512_permutexvar_epi32(circ8, v_v);
-            __m512i v_v9 = _mm512_permutexvar_epi32(circ9, v_v);
-            __m512i v_v10 = _mm512_permutexvar_epi32(circ10, v_v);
-            __m512i v_v11 = _mm512_permutexvar_epi32(circ11, v_v);
-            __m512i v_v12 = _mm512_permutexvar_epi32(circ12, v_v);
-            __m512i v_v13 = _mm512_permutexvar_epi32(circ13, v_v);
-            __m512i v_v14 = _mm512_permutexvar_epi32(circ14, v_v);
-            __m512i v_v15 = _mm512_permutexvar_epi32(circ15, v_v);
-            __mmask16 tmp_match1 = _mm512_cmpeq_epi32_mask(v_u, v_v1); // find matches
-            __mmask16 tmp_match2 = _mm512_cmpeq_epi32_mask(v_u, v_v2);
-            __mmask16 tmp_match3 = _mm512_cmpeq_epi32_mask(v_u, v_v3);
-            __mmask16 tmp_match4 = _mm512_cmpeq_epi32_mask(v_u, v_v4);
-            __mmask16 tmp_match5 = _mm512_cmpeq_epi32_mask(v_u, v_v5);
-            __mmask16 tmp_match6 = _mm512_cmpeq_epi32_mask(v_u, v_v6);
-            __mmask16 tmp_match7 = _mm512_cmpeq_epi32_mask(v_u, v_v7);
-            __mmask16 tmp_match8 = _mm512_cmpeq_epi32_mask(v_u, v_v8);
-            __mmask16 tmp_match9 = _mm512_cmpeq_epi32_mask(v_u, v_v9);
-            __mmask16 tmp_match10 = _mm512_cmpeq_epi32_mask(v_u, v_v10);
-            __mmask16 tmp_match11 = _mm512_cmpeq_epi32_mask(v_u, v_v11);
-            __mmask16 tmp_match12 = _mm512_cmpeq_epi32_mask(v_u, v_v12);
-            __mmask16 tmp_match13 = _mm512_cmpeq_epi32_mask(v_u, v_v13);
-            __mmask16 tmp_match14 = _mm512_cmpeq_epi32_mask(v_u, v_v14);
-            __mmask16 tmp_match15 = _mm512_cmpeq_epi32_mask(v_u, v_v15);
-            match = _mm512_kor(
-                _mm512_kor(
-                    _mm512_kor(_mm512_kor(match, tmp_match1), _mm512_kor(tmp_match2, tmp_match3)),
-                    _mm512_kor(_mm512_kor(tmp_match4, tmp_match5),
-                               _mm512_kor(tmp_match6, tmp_match7))),
-                _mm512_kor(
-                    _mm512_kor(_mm512_kor(tmp_match8, tmp_match9),
-                               _mm512_kor(tmp_match10, tmp_match11)),
-                    _mm512_kor(_mm512_kor(tmp_match12, tmp_match13),
-                               _mm512_kor(tmp_match14, tmp_match15)))); // combine all matches
-        }
-        total += _popcnt32_redef(_mm512_mask2int(match)); //count number of matches
-    }
-
-    while (i_u < (n_u / 16) * 16 && i_v < n_v) {
-        __m512i v_u = _mm512_loadu_si512((void *)(neigh_u + i_u));
-        while (neigh_v[i_v] <= neigh_u[i_u + 15] && i_v < n_v) {
-            __m512i tmp_v_v = _mm512_set1_epi32(neigh_v[i_v]);
-            __mmask16 match = _mm512_cmpeq_epi32_mask(v_u, tmp_v_v);
-            if (_mm512_mask2int(match))
-                total++;
-            i_v++;
-        }
-        i_u += 16;
-    }
-    while (i_v < (n_v / 16) * 16 && i_u < n_u) {
-        __m512i v_v = _mm512_loadu_si512((void *)(neigh_v + i_v));
-        while (neigh_u[i_u] <= neigh_v[i_v + 15] && i_u < n_u) {
-            __m512i tmp_v_u = _mm512_set1_epi32(neigh_u[i_u]);
-            __mmask16 match = _mm512_cmpeq_epi32_mask(v_v, tmp_v_u);
-            if (_mm512_mask2int(match))
-                total++;
-            i_u++;
-        }
-        i_v += 16;
-    }
-
-    while (i_u <= (n_u - 8) && i_v <= (n_v - 8)) { // not in last n%8 elements
-        // assumes neighbor list is ordered
-        std::int32_t min_neigh_u = neigh_u[i_u];
-        std::int32_t max_neigh_v = neigh_v[i_v + 7];
-
-        if (min_neigh_u > max_neigh_v) {
-            if (min_neigh_u > neigh_v[n_v - 1]) {
-                return total;
-            }
-            i_v += 8;
-            continue;
-        }
-        std::int32_t max_neigh_u = neigh_u[i_u + 7];
-        std::int32_t min_neigh_v = neigh_v[i_v];
-        if (min_neigh_v > max_neigh_u) {
-            if (min_neigh_v > neigh_u[n_u - 1]) {
-                return total;
-            }
-            i_u += 8;
-            continue;
-        }
-        __m256i v_u = _mm256_loadu_si256(
-            reinterpret_cast<const __m256i *>(neigh_u + i_u)); // load 8 neighbors of u
-        __m256i v_v = _mm256_loadu_si256(
-            reinterpret_cast<const __m256i *>(neigh_v + i_v)); // load 8 neighbors of v
-
-        if (max_neigh_u >= max_neigh_v)
-            i_v += 8;
-        if (max_neigh_u <= max_neigh_v)
-            i_u += 8;
-
-        __mmask8 match = _mm256_cmpeq_epi32_mask(v_u, v_v);
-        if (_cvtmask8_u32(match) != 0xff) { // shortcut case where all neighbors match
-            __m256i circ1 = _mm256_set_epi32(0, 7, 6, 5, 4, 3, 2, 1);
-            __m256i circ2 = _mm256_set_epi32(1, 0, 7, 6, 5, 4, 3, 2);
-            __m256i circ3 = _mm256_set_epi32(2, 1, 0, 7, 6, 5, 4, 3);
-            __m256i circ4 = _mm256_set_epi32(3, 2, 1, 0, 7, 6, 5, 4);
-            __m256i circ5 = _mm256_set_epi32(4, 3, 2, 1, 0, 7, 6, 5);
-            __m256i circ6 = _mm256_set_epi32(5, 4, 3, 2, 1, 0, 7, 6);
-            __m256i circ7 = _mm256_set_epi32(6, 5, 4, 3, 2, 1, 0, 7);
-
-            __m256i v_v1 = _mm256_permutexvar_epi32(circ1, v_v);
-            __m256i v_v2 = _mm256_permutexvar_epi32(circ2, v_v);
-            __m256i v_v3 = _mm256_permutexvar_epi32(circ3, v_v);
-            __m256i v_v4 = _mm256_permutexvar_epi32(circ4, v_v);
-            __m256i v_v5 = _mm256_permutexvar_epi32(circ5, v_v);
-            __m256i v_v6 = _mm256_permutexvar_epi32(circ6, v_v);
-            __m256i v_v7 = _mm256_permutexvar_epi32(circ7, v_v);
-
-            __mmask8 tmp_match1 = _mm256_cmpeq_epi32_mask(v_u, v_v1); // find matches
-            __mmask8 tmp_match2 = _mm256_cmpeq_epi32_mask(v_u, v_v2);
-            __mmask8 tmp_match3 = _mm256_cmpeq_epi32_mask(v_u, v_v3);
-            __mmask8 tmp_match4 = _mm256_cmpeq_epi32_mask(v_u, v_v4);
-            __mmask8 tmp_match5 = _mm256_cmpeq_epi32_mask(v_u, v_v5);
-            __mmask8 tmp_match6 = _mm256_cmpeq_epi32_mask(v_u, v_v6);
-            __mmask8 tmp_match7 = _mm256_cmpeq_epi32_mask(v_u, v_v7);
-
-            match = _kor_mask8(
-                _kor_mask8(_kor_mask8(match, tmp_match1), _kor_mask8(tmp_match2, tmp_match3)),
-                _kor_mask8(_kor_mask8(tmp_match4, tmp_match5),
-                           _kor_mask8(tmp_match6, tmp_match7))); // combine all matches
-        }
-        total += _popcnt32_redef(_cvtmask8_u32(match)); //count number of matches
-    }
-    if (i_u <= (n_u - 8) && i_v < n_v) {
-        __m256i v_u = _mm256_loadu_si256(reinterpret_cast<const __m256i *>(neigh_u + i_u));
-        while (neigh_v[i_v] <= neigh_u[i_u + 7] && i_v < n_v) {
-            __m256i tmp_v_v = _mm256_set1_epi32(neigh_v[i_v]);
-            __mmask8 match = _mm256_cmpeq_epi32_mask(v_u, tmp_v_v);
-            if (_cvtmask8_u32(match))
-                total++;
-            i_v++;
-        }
-        i_u += 8;
-    }
-    if (i_v <= (n_v - 8) && i_u < n_u) {
-        __m256i v_v = _mm256_loadu_si256(reinterpret_cast<const __m256i *>(neigh_v + i_v));
-        while (neigh_u[i_u] <= neigh_v[i_v + 7] && i_u < n_u) {
-            __m256i tmp_v_u = _mm256_set1_epi32(neigh_u[i_u]);
-            __mmask8 match = _mm256_cmpeq_epi32_mask(v_v, tmp_v_u);
-            if (_cvtmask8_u32(match))
-                total++;
-            i_u++;
-        }
-        i_v += 8;
-    }
-
-    while (i_u <= (n_u - 4) && i_v <= (n_v - 4)) { // not in last n%8 elements
-        // assumes neighbor list is ordered
-        std::int32_t min_neigh_u = neigh_u[i_u];
-        std::int32_t max_neigh_v = neigh_v[i_v + 3];
-
-        if (min_neigh_u > max_neigh_v) {
-            if (min_neigh_u > neigh_v[n_v - 1]) {
-                return total;
-            }
-            i_v += 4;
-            continue;
-        }
-        std::int32_t min_neigh_v = neigh_v[i_v];
-        std::int32_t max_neigh_u = neigh_u[i_u + 3];
-        if (min_neigh_v > max_neigh_u) {
-            if (min_neigh_v > neigh_u[n_u - 1]) {
-                return total;
-            }
-            i_u += 4;
-            continue;
-        }
-        __m128i v_u = _mm_loadu_si128(
-            reinterpret_cast<const __m128i *>(neigh_u + i_u)); // load 8 neighbors of u
-        __m128i v_v = _mm_loadu_si128(
-            reinterpret_cast<const __m128i *>(neigh_v + i_v)); // load 8 neighbors of v
-
-        if (max_neigh_u >= max_neigh_v)
-            i_v += 4;
-        if (max_neigh_u <= max_neigh_v)
-            i_u += 4;
-
-        __mmask8 match = _mm_cmpeq_epi32_mask(v_u, v_v);
-        if (_cvtmask8_u32(match) != 0xf) { // shortcut case where all neighbors match
-            __m128i v_v1 = _mm_shuffle_epi32(v_v, _MM_SHUFFLE(0, 3, 2, 1));
-            __m128i v_v2 = _mm_shuffle_epi32(v_v, _MM_SHUFFLE(1, 0, 3, 2));
-            __m128i v_v3 = _mm_shuffle_epi32(v_v, _MM_SHUFFLE(2, 1, 0, 3));
-
-            __mmask8 tmp_match1 = _mm_cmpeq_epi32_mask(v_u, v_v1); // find matches
-            __mmask8 tmp_match2 = _mm_cmpeq_epi32_mask(v_u, v_v2);
-            __mmask8 tmp_match3 = _mm_cmpeq_epi32_mask(v_u, v_v3);
-
-            match = _kor_mask8(_kor_mask8(match, tmp_match1),
-                               _kor_mask8(tmp_match2, tmp_match3)); // combine all matches
-        }
-        total += _popcnt32_redef(_cvtmask8_u32(match)); //count number of matches
-    }
-    if (i_u <= (n_u - 4) && i_v < n_v) {
-        __m128i v_u = _mm_loadu_si128(reinterpret_cast<const __m128i *>(neigh_u + i_u));
-        while (neigh_v[i_v] <= neigh_u[i_u + 3] && i_v < n_v) {
-            __m128i tmp_v_v = _mm_set1_epi32(neigh_v[i_v]);
-            __mmask8 match = _mm_cmpeq_epi32_mask(v_u, tmp_v_v);
-            if (_cvtmask8_u32(match))
-                total++;
-            i_v++;
-        }
-        i_u += 4;
-    }
-    if (i_v <= (n_v - 4) && i_u < n_u) {
-        __m128i v_v = _mm_loadu_si128(reinterpret_cast<const __m128i *>(neigh_v + i_v));
-        while (neigh_u[i_u] <= neigh_v[i_v + 3] && i_u < n_u) {
-            __m128i tmp_v_u = _mm_set1_epi32(neigh_u[i_u]);
-            __mmask8 match = _mm_cmpeq_epi32_mask(v_v, tmp_v_u);
-            if (_cvtmask8_u32(match))
-                total++;
-            i_u++;
-        }
-        i_v += 4;
-    }
-#endif
-    while (i_u < n_u && i_v < n_v) {
-        if ((neigh_u[i_u] > neigh_v[n_v - 1]) || (neigh_v[i_v] > neigh_u[n_u - 1])) {
-            return total;
-        }
-        if (neigh_u[i_u] == neigh_v[i_v])
-            total++, i_u++, i_v++;
-        else if (neigh_u[i_u] < neigh_v[i_v])
-            i_u++;
-        else if (neigh_u[i_u] > neigh_v[i_v])
-            i_v++;
-    }
-    return total;
-}
+using namespace preview::backend;
 
 template <typename Cpu>
-vertex_similarity_result call_jaccard_default_kernel_avx512(
-    const descriptor_base &desc,
-    const dal::preview::detail::topology<std::int32_t> &data,
+vertex_similarity_result<task::all_vertex_pairs> jaccard_avx512(
+    const detail::descriptor_base<task::all_vertex_pairs> &desc,
+    const dal::preview::detail::topology<std::int32_t> &t,
     void *result_ptr) {
-    const auto g_edge_offsets = data._rows_vertex.get_data();
-    const auto g_vertex_neighbors = data._cols.get_data();
-    const auto g_degrees = data._degrees.get_data();
+    const auto rows_vertex = t._rows_vertex.get_data();
+    const auto cols = t._cols.get_data();
+    const auto degrees = t._degrees.get_data();
 
     const auto row_begin = dal::detail::integral_cast<std::int32_t>(desc.get_row_range_begin());
     const auto row_end = dal::detail::integral_cast<std::int32_t>(desc.get_row_range_end());
@@ -358,7 +52,7 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
         dal::detail::integral_cast<std::int32_t>(desc.get_column_range_begin());
     const auto column_end = dal::detail::integral_cast<std::int32_t>(desc.get_column_range_end());
     const auto number_elements_in_block =
-        compute_number_elements_in_block(row_begin, row_end, column_begin, column_end);
+        detail::compute_number_elements_in_block(row_begin, row_end, column_begin, column_end);
     std::int32_t *first_vertices = reinterpret_cast<std::int32_t *>(result_ptr);
     std::int32_t *second_vertices = first_vertices + number_elements_in_block;
     float *jaccard = reinterpret_cast<float *>(second_vertices + number_elements_in_block);
@@ -374,9 +68,9 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 #endif
 
     for (std::int32_t i = row_begin; i < row_end; ++i) {
-        const auto i_neighbor_size = g_degrees[i];
-        const auto i_neigbhors = g_vertex_neighbors + g_edge_offsets[i];
-        const auto diagonal = min(i, column_end);
+        const auto i_neighbor_size = degrees[i];
+        const auto i_neigbhors = cols + rows_vertex[i];
+        const auto diagonal = detail::min(i, column_end);
 
 #if defined(__INTEL_COMPILER)
         __m512i n_i_start_v = _mm512_set1_epi32(i_neigbhors[0]);
@@ -385,29 +79,29 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 
         if (j < column_begin + ((diagonal - column_begin) / 16) * 16) {
             //load_data(0)
-            __m512i start_indices_j_v = _mm512_load_epi32(g_edge_offsets + j);
-            __m512i end_indices_j_v_tmp = _mm512_load_epi32(g_edge_offsets + j + 1);
+            __m512i start_indices_j_v = _mm512_load_epi32(rows_vertex + j);
+            __m512i end_indices_j_v_tmp = _mm512_load_epi32(rows_vertex + j + 1);
             __m512i end_indices_j_v = _mm512_add_epi32(end_indices_j_v_tmp, _mm512_set1_epi32(-1));
 
-            __m512i n_j_start_v = _mm512_permutexvar_epi32(
-                j_vertices_tmp1,
-                _mm512_i32gather_epi32(start_indices_j_v, g_vertex_neighbors, 4));
-            __m512i n_j_end_v = _mm512_permutexvar_epi32(
-                j_vertices_tmp1,
-                _mm512_i32gather_epi32(end_indices_j_v, g_vertex_neighbors, 4));
+            __m512i n_j_start_v =
+                _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                         _mm512_i32gather_epi32(start_indices_j_v, cols, 4));
+            __m512i n_j_end_v =
+                _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                         _mm512_i32gather_epi32(end_indices_j_v, cols, 4));
 
             for (; j + 16 < column_begin + ((diagonal - column_begin) / 16) * 16;) {
-                __m512i start_indices_j_v = _mm512_load_epi32(g_edge_offsets + j + 16);
-                __m512i end_indices_j_v_tmp = _mm512_load_epi32(g_edge_offsets + j + 17);
+                __m512i start_indices_j_v = _mm512_load_epi32(rows_vertex + j + 16);
+                __m512i end_indices_j_v_tmp = _mm512_load_epi32(rows_vertex + j + 17);
                 __m512i end_indices_j_v =
                     _mm512_add_epi32(end_indices_j_v_tmp, _mm512_set1_epi32(-1));
 
-                __m512i n_j_start_v1 = _mm512_permutexvar_epi32(
-                    j_vertices_tmp1,
-                    _mm512_i32gather_epi32(start_indices_j_v, g_vertex_neighbors, 4));
-                __m512i n_j_end_v1 = _mm512_permutexvar_epi32(
-                    j_vertices_tmp1,
-                    _mm512_i32gather_epi32(end_indices_j_v, g_vertex_neighbors, 4));
+                __m512i n_j_start_v1 =
+                    _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                             _mm512_i32gather_epi32(start_indices_j_v, cols, 4));
+                __m512i n_j_end_v1 =
+                    _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                             _mm512_i32gather_epi32(end_indices_j_v, cols, 4));
 
                 __mmask16 cmpgt1 = _mm512_cmpgt_epi32_mask(n_i_start_v, n_j_end_v);
                 __mmask16 cmpgt2 = _mm512_cmpgt_epi32_mask(n_j_start_v, n_i_end_v);
@@ -425,10 +119,10 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 
                     GRAPH_STACK_ALING(64) std::int32_t stack16_intersections[16] = { 0 };
                     for (std::int32_t s = 0; s < ones_num; s++) {
-                        const auto j_neighbor_size = g_degrees[stack16_j_vertex[s]];
-                        const auto j_neigbhors =
-                            g_vertex_neighbors + g_edge_offsets[stack16_j_vertex[s]];
-                        stack16_intersections[s] = intersection(i_neigbhors,
+                        const auto j_neighbor_size = degrees[stack16_j_vertex[s]];
+                        const auto j_neigbhors = cols + rows_vertex[stack16_j_vertex[s]];
+                        stack16_intersections[s] =
+                            preview::backend::intersection<Cpu>(i_neigbhors,
                                                                 j_neigbhors,
                                                                 i_neighbor_size,
                                                                 j_neighbor_size);
@@ -475,11 +169,12 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 
                 GRAPH_STACK_ALING(64) std::int32_t stack16_intersections[16] = { 0 };
                 for (std::int32_t s = 0; s < ones_num; s++) {
-                    const auto j_neighbor_size = g_degrees[stack16_j_vertex[s]];
-                    const auto j_neigbhors =
-                        g_vertex_neighbors + g_edge_offsets[stack16_j_vertex[s]];
-                    stack16_intersections[s] =
-                        intersection(i_neigbhors, j_neigbhors, i_neighbor_size, j_neighbor_size);
+                    const auto j_neighbor_size = degrees[stack16_j_vertex[s]];
+                    const auto j_neigbhors = cols + rows_vertex[stack16_j_vertex[s]];
+                    stack16_intersections[s] = preview::backend::intersection<Cpu>(i_neigbhors,
+                                                                                   j_neigbhors,
+                                                                                   i_neighbor_size,
+                                                                                   j_neighbor_size);
                 }
                 __m512i intersections_v = _mm512_load_epi32(stack16_intersections);
                 j_vertices = _mm512_load_epi32(stack16_j_vertex);
@@ -501,12 +196,14 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
             j += 16;
 
             for (j = column_begin + ((diagonal - column_begin) / 16); j < diagonal; j++) {
-                const auto j_neighbor_size = g_degrees[j];
-                const auto j_neigbhors = g_vertex_neighbors + g_edge_offsets[j];
+                const auto j_neighbor_size = degrees[j];
+                const auto j_neigbhors = cols + rows_vertex[j];
                 if (!(i_neigbhors[0] > j_neigbhors[j_neighbor_size - 1]) &&
                     !(j_neigbhors[0] > i_neigbhors[i_neighbor_size - 1])) {
-                    auto intersection_value =
-                        intersection(i_neigbhors, j_neigbhors, i_neighbor_size, j_neighbor_size);
+                    auto intersection_value = preview::backend::intersection<Cpu>(i_neigbhors,
+                                                                                  j_neigbhors,
+                                                                                  i_neighbor_size,
+                                                                                  j_neighbor_size);
                     if (intersection_value) {
                         jaccard[nnz] = static_cast<float>(intersection_value);
                         first_vertices[nnz] = i;
@@ -520,12 +217,14 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
         else {
 #endif
             for (j = column_begin; j < diagonal; j++) {
-                const auto j_neighbor_size = g_degrees[j];
-                const auto j_neigbhors = g_vertex_neighbors + g_edge_offsets[j];
+                const auto j_neighbor_size = degrees[j];
+                const auto j_neigbhors = cols + rows_vertex[j];
                 if (!(i_neigbhors[0] > j_neigbhors[j_neighbor_size - 1]) &&
                     !(j_neigbhors[0] > i_neigbhors[i_neighbor_size - 1])) {
-                    auto intersection_value =
-                        intersection(i_neigbhors, j_neigbhors, i_neighbor_size, j_neighbor_size);
+                    auto intersection_value = preview::backend::intersection<Cpu>(i_neigbhors,
+                                                                                  j_neigbhors,
+                                                                                  i_neighbor_size,
+                                                                                  j_neighbor_size);
                     if (intersection_value) {
                         jaccard[nnz] = static_cast<float>(intersection_value);
                         first_vertices[nnz] = i;
@@ -553,29 +252,29 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 #if defined(__INTEL_COMPILER)
         if (j < tmp_idx + ((column_end - tmp_idx) / 16) * 16) {
             //load_data(0)
-            __m512i start_indices_j_v = _mm512_load_epi32(g_edge_offsets + j);
-            __m512i end_indices_j_v_tmp = _mm512_load_epi32(g_edge_offsets + j + 1);
+            __m512i start_indices_j_v = _mm512_load_epi32(rows_vertex + j);
+            __m512i end_indices_j_v_tmp = _mm512_load_epi32(rows_vertex + j + 1);
             __m512i end_indices_j_v = _mm512_add_epi32(end_indices_j_v_tmp, _mm512_set1_epi32(-1));
 
-            __m512i n_j_start_v = _mm512_permutexvar_epi32(
-                j_vertices_tmp1,
-                _mm512_i32gather_epi32(start_indices_j_v, g_vertex_neighbors, 4));
-            __m512i n_j_end_v = _mm512_permutexvar_epi32(
-                j_vertices_tmp1,
-                _mm512_i32gather_epi32(end_indices_j_v, g_vertex_neighbors, 4));
+            __m512i n_j_start_v =
+                _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                         _mm512_i32gather_epi32(start_indices_j_v, cols, 4));
+            __m512i n_j_end_v =
+                _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                         _mm512_i32gather_epi32(end_indices_j_v, cols, 4));
 
             for (; j + 16 < tmp_idx + ((column_end - tmp_idx) / 16) * 16;) {
-                __m512i start_indices_j_v = _mm512_load_epi32(g_edge_offsets + j + 16);
-                __m512i end_indices_j_v_tmp = _mm512_load_epi32(g_edge_offsets + j + 17);
+                __m512i start_indices_j_v = _mm512_load_epi32(rows_vertex + j + 16);
+                __m512i end_indices_j_v_tmp = _mm512_load_epi32(rows_vertex + j + 17);
                 __m512i end_indices_j_v =
                     _mm512_add_epi32(end_indices_j_v_tmp, _mm512_set1_epi32(-1));
 
-                __m512i n_j_start_v1 = _mm512_permutexvar_epi32(
-                    j_vertices_tmp1,
-                    _mm512_i32gather_epi32(start_indices_j_v, g_vertex_neighbors, 4));
-                __m512i n_j_end_v1 = _mm512_permutexvar_epi32(
-                    j_vertices_tmp1,
-                    _mm512_i32gather_epi32(end_indices_j_v, g_vertex_neighbors, 4));
+                __m512i n_j_start_v1 =
+                    _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                             _mm512_i32gather_epi32(start_indices_j_v, cols, 4));
+                __m512i n_j_end_v1 =
+                    _mm512_permutexvar_epi32(j_vertices_tmp1,
+                                             _mm512_i32gather_epi32(end_indices_j_v, cols, 4));
 
                 __mmask16 cmpgt1 = _mm512_cmpgt_epi32_mask(n_i_start_v, n_j_end_v);
                 __mmask16 cmpgt2 = _mm512_cmpgt_epi32_mask(n_j_start_v, n_i_end_v);
@@ -593,10 +292,10 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 
                     GRAPH_STACK_ALING(64) std::int32_t stack16_intersections[16] = { 0 };
                     for (std::int32_t s = 0; s < ones_num; s++) {
-                        const auto j_neighbor_size = g_degrees[stack16_j_vertex[s]];
-                        const auto j_neigbhors =
-                            g_vertex_neighbors + g_edge_offsets[stack16_j_vertex[s]];
-                        stack16_intersections[s] = intersection(i_neigbhors,
+                        const auto j_neighbor_size = degrees[stack16_j_vertex[s]];
+                        const auto j_neigbhors = cols + rows_vertex[stack16_j_vertex[s]];
+                        stack16_intersections[s] =
+                            preview::backend::intersection<Cpu>(i_neigbhors,
                                                                 j_neigbhors,
                                                                 i_neighbor_size,
                                                                 j_neighbor_size);
@@ -643,11 +342,12 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
 
                 GRAPH_STACK_ALING(64) std::int32_t stack16_intersections[16] = { 0 };
                 for (std::int32_t s = 0; s < ones_num; s++) {
-                    const auto j_neighbor_size = g_degrees[stack16_j_vertex[s]];
-                    const auto j_neigbhors =
-                        g_vertex_neighbors + g_edge_offsets[stack16_j_vertex[s]];
-                    stack16_intersections[s] =
-                        intersection(i_neigbhors, j_neigbhors, i_neighbor_size, j_neighbor_size);
+                    const auto j_neighbor_size = degrees[stack16_j_vertex[s]];
+                    const auto j_neigbhors = cols + rows_vertex[stack16_j_vertex[s]];
+                    stack16_intersections[s] = preview::backend::intersection<Cpu>(i_neigbhors,
+                                                                                   j_neigbhors,
+                                                                                   i_neighbor_size,
+                                                                                   j_neighbor_size);
                 }
                 __m512i intersections_v = _mm512_load_epi32(stack16_intersections);
                 j_vertices = _mm512_load_epi32(stack16_j_vertex);
@@ -669,12 +369,14 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
             j += 16;
 
             for (j = tmp_idx + ((column_end - tmp_idx) / 16) * 16; j < column_end; j++) {
-                const auto j_neighbor_size = g_degrees[j];
-                const auto j_neigbhors = g_vertex_neighbors + g_edge_offsets[j];
+                const auto j_neighbor_size = degrees[j];
+                const auto j_neigbhors = cols + rows_vertex[j];
                 if (!(i_neigbhors[0] > j_neigbhors[j_neighbor_size - 1]) &&
                     !(j_neigbhors[0] > i_neigbhors[i_neighbor_size - 1])) {
-                    auto intersection_value =
-                        intersection(i_neigbhors, j_neigbhors, i_neighbor_size, j_neighbor_size);
+                    auto intersection_value = preview::backend::intersection<Cpu>(i_neigbhors,
+                                                                                  j_neigbhors,
+                                                                                  i_neighbor_size,
+                                                                                  j_neighbor_size);
                     if (intersection_value) {
                         jaccard[nnz] = static_cast<float>(intersection_value);
                         first_vertices[nnz] = i;
@@ -688,12 +390,14 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
         else {
 #endif
             for (j = tmp_idx; j < column_end; j++) {
-                const auto j_neighbor_size = g_degrees[j];
-                const auto j_neigbhors = g_vertex_neighbors + g_edge_offsets[j];
+                const auto j_neighbor_size = degrees[j];
+                const auto j_neigbhors = cols + rows_vertex[j];
                 if (!(i_neigbhors[0] > j_neigbhors[j_neighbor_size - 1]) &&
                     !(j_neigbhors[0] > i_neigbhors[i_neighbor_size - 1])) {
-                    auto intersection_value =
-                        intersection(i_neigbhors, j_neigbhors, i_neighbor_size, j_neighbor_size);
+                    auto intersection_value = preview::backend::intersection<Cpu>(i_neigbhors,
+                                                                                  j_neigbhors,
+                                                                                  i_neighbor_size,
+                                                                                  j_neighbor_size);
                     if (intersection_value) {
                         jaccard[nnz] = static_cast<float>(intersection_value);
                         first_vertices[nnz] = i;
@@ -711,9 +415,8 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
     PRAGMA_VECTOR_ALWAYS
     for (int i = 0; i < nnz; i++) {
         if (first_vertices[i] != second_vertices[i])
-            jaccard[i] =
-                jaccard[i] / static_cast<float>(g_degrees[first_vertices[i]] +
-                                                g_degrees[second_vertices[i]] - jaccard[i]);
+            jaccard[i] = jaccard[i] / static_cast<float>(degrees[first_vertices[i]] +
+                                                         degrees[second_vertices[i]] - jaccard[i]);
     }
 
     vertex_similarity_result res(
@@ -722,6 +425,5 @@ vertex_similarity_result call_jaccard_default_kernel_avx512(
         nnz);
     return res;
 }
-} // namespace detail
-} // namespace jaccard
-} // namespace oneapi::dal::preview
+
+} // namespace oneapi::dal::preview::jaccard::backend
