+++ cpp/oneapi/dal/algo/jaccard/detail/vertex_similarity_default_kernel.cpp
@@ -14,30 +14,26 @@
 * limitations under the License.
 *******************************************************************************/
 
-#include "oneapi/dal/algo/jaccard/detail/select_kernel.hpp"
 #include "oneapi/dal/algo/jaccard/backend/cpu/vertex_similarity_default_kernel.hpp"
+#include "oneapi/dal/algo/jaccard/detail/vertex_similarity_default_kernel.hpp"
 #include "oneapi/dal/detail/policy.hpp"
 #include "oneapi/dal/backend/dispatcher.hpp"
 
 namespace oneapi::dal::preview::jaccard::detail {
 
-template <typename Float, typename Method>
-vertex_similarity_result backend_default<dal::detail::host_policy,
-                                         Float,
-                                         Method,
-                                         dal::preview::detail::topology<std::int32_t>>::
-operator()(const dal::detail::host_policy &policy,
-           const descriptor_base &desc,
-           const dal::preview::detail::topology<std::int32_t> &data,
-           void *result_ptr) {
-    return dal::backend::dispatch_by_cpu(dal::backend::context_cpu{ policy }, [&](auto cpu) {
-        return call_jaccard_default_kernel_int32<decltype(cpu)>(desc, data, result_ptr);
+template <typename Float>
+vertex_similarity_result<task::all_vertex_pairs>
+vertex_similarity<Float, task::all_vertex_pairs, dal::preview::detail::topology<std::int32_t>>::
+operator()(const dal::detail::host_policy& ctx,
+           const detail::descriptor_base<task::all_vertex_pairs>& desc,
+           const dal::preview::detail::topology<std::int32_t>& t,
+           void* result_ptr) {
+    return dal::backend::dispatch_by_cpu(dal::backend::context_cpu{ ctx }, [&](auto cpu) {
+        return backend::jaccard<decltype(cpu)>(desc, t, result_ptr);
     });
 }
 
-template struct ONEDAL_EXPORT backend_default<dal::detail::host_policy,
-                                              float,
-                                              dal::preview::jaccard::method::fast,
-                                              dal::preview::detail::topology<std::int32_t>>;
+template struct ONEDAL_EXPORT
+    vertex_similarity<float, task::all_vertex_pairs, dal::preview::detail::topology<std::int32_t>>;
 
 } // namespace oneapi::dal::preview::jaccard::detail
