+++ cpp/oneapi/dal/algo/jaccard/vertex_similarity_types.cpp
@@ -16,8 +16,7 @@
 
 #include "oneapi/dal/algo/jaccard/vertex_similarity_types.hpp"
 
-namespace oneapi::dal::preview {
-namespace jaccard {
+namespace oneapi::dal::preview::jaccard {
 
 class detail::vertex_similarity_result_impl : public base {
 public:
@@ -28,27 +27,35 @@ public:
 
 using detail::vertex_similarity_result_impl;
 
-vertex_similarity_result::vertex_similarity_result() : impl_(new vertex_similarity_result_impl()) {}
+template <typename Task>
+vertex_similarity_result<Task>::vertex_similarity_result()
+        : impl_(new vertex_similarity_result_impl()) {}
 
-vertex_similarity_result::vertex_similarity_result(const table& vertex_pairs,
-                                                   const table& coeffs,
-                                                   std::int64_t nonzero_coeff_count)
+template <typename Task>
+vertex_similarity_result<Task>::vertex_similarity_result(const table& vertex_pairs,
+                                                         const table& coeffs,
+                                                         std::int64_t nonzero_coeff_count)
         : impl_(new vertex_similarity_result_impl()) {
     impl_->vertex_pairs = vertex_pairs;
     impl_->coeffs = coeffs;
     impl_->nonzero_coeff_count = nonzero_coeff_count;
 }
 
-table vertex_similarity_result::get_coeffs() const {
+template <typename Task>
+const table& vertex_similarity_result<Task>::get_coeffs() const {
     return impl_->coeffs;
 }
 
-table vertex_similarity_result::get_vertex_pairs() const {
+template <typename Task>
+const table& vertex_similarity_result<Task>::get_vertex_pairs() const {
     return impl_->vertex_pairs;
 }
 
-int64_t vertex_similarity_result::get_nonzero_coeff_count() const {
+template <typename Task>
+int64_t vertex_similarity_result<Task>::get_nonzero_coeff_count() const {
     return impl_->nonzero_coeff_count;
 }
-} // namespace jaccard
-} // namespace oneapi::dal::preview
+
+template class ONEDAL_EXPORT vertex_similarity_result<task::all_vertex_pairs>;
+
+} // namespace oneapi::dal::preview::jaccard
