+++ cpp/oneapi/dal/algo/jaccard/vertex_similarity_types.hpp
@@ -23,23 +23,25 @@
 #include "oneapi/dal/algo/jaccard/common.hpp"
 #include "oneapi/dal/algo/jaccard/detail/vertex_similarity_types.hpp"
 
-namespace oneapi::dal::preview {
-namespace jaccard {
+namespace oneapi::dal::preview::jaccard {
 
 /// Class for the description of the input parameters of the Jaccard Similarity
 /// algorithm
 ///
 /// @tparam Graph  Type of the input graph
-template <typename Graph>
-class ONEDAL_EXPORT vertex_similarity_input {
+template <typename Graph, typename Task = task::by_default>
+class vertex_similarity_input : public base {
+    static_assert(detail::is_valid_task<Task>);
+
 public:
+    using task_t = Task;
     static_assert(detail::is_valid_graph<Graph>,
                   "Only undirected_adjacency_vector_graph is supported.");
     /// Constructs the algorithm input initialized with the graph and the caching builder.
     ///
     /// @param [in]   graph  The input graph
     /// @param [in/out]  builder  The caching builder
-    vertex_similarity_input(const Graph& graph, caching_builder& builder);
+    vertex_similarity_input(const Graph& g, caching_builder& builder);
 
     /// Returns the constant reference to the input graph
     const Graph& get_graph() const;
@@ -48,12 +50,16 @@ public:
     caching_builder& get_caching_builder();
 
 private:
-    dal::detail::pimpl<detail::vertex_similarity_input_impl<Graph>> impl_;
+    dal::detail::pimpl<detail::vertex_similarity_input_impl<Graph, Task>> impl_;
 };
 
 /// Class for the description of the result of the Jaccard Similarity algorithm
-class ONEDAL_EXPORT vertex_similarity_result {
+template <typename Task = task::by_default>
+class vertex_similarity_result {
+    static_assert(detail::is_valid_task<Task>);
+
 public:
+    using task_t = Task;
     /// Constructs the empty result
     vertex_similarity_result();
 
@@ -73,11 +79,11 @@ public:
 
     /// Returns the table of size [nonzero_coeff_count x 1] with non-zero Jaccard
     /// similarity coefficients
-    table get_coeffs() const;
+    const table& get_coeffs() const;
 
     /// Returns the table of size [nonzero_coeff_count x 2] with vertex pairs which have
     /// non-zero Jaccard similarity coefficients
-    table get_vertex_pairs() const;
+    const table& get_vertex_pairs() const;
 
     /// The number of non-zero Jaccard similarity coefficients in the block
     std::int64_t get_nonzero_coeff_count() const;
@@ -86,20 +92,19 @@ private:
     dal::detail::pimpl<detail::vertex_similarity_result_impl> impl_;
 };
 
-template <typename Graph>
-vertex_similarity_input<Graph>::vertex_similarity_input(const Graph& data,
-                                                        caching_builder& builder_input)
-        : impl_(new detail::vertex_similarity_input_impl<Graph>(data, builder_input)) {}
+template <typename Graph, typename Task>
+vertex_similarity_input<Graph, Task>::vertex_similarity_input(const Graph& data,
+                                                              caching_builder& builder_input)
+        : impl_(new detail::vertex_similarity_input_impl<Graph, Task>(data, builder_input)) {}
 
-template <typename Graph>
-const Graph& vertex_similarity_input<Graph>::get_graph() const {
+template <typename Graph, typename Task>
+const Graph& vertex_similarity_input<Graph, Task>::get_graph() const {
     return impl_->graph_data;
 }
 
-template <typename Graph>
-caching_builder& vertex_similarity_input<Graph>::get_caching_builder() {
+template <typename Graph, typename Task>
+caching_builder& vertex_similarity_input<Graph, Task>::get_caching_builder() {
     return impl_->builder;
 }
 
-} // namespace jaccard
-} // namespace oneapi::dal::preview
+} // namespace oneapi::dal::preview::jaccard
