+++ cpp/oneapi/dal/algo/triangle_counting/detail/vertex_ranking_default_kernel.hpp
@@ -17,6 +17,7 @@
 #pragma once
 
 #include "oneapi/dal/algo/triangle_counting/common.hpp"
+#include "oneapi/dal/algo/triangle_counting/detail/relabel_kernel.hpp"
 #include "oneapi/dal/algo/triangle_counting/vertex_ranking_types.hpp"
 #include "oneapi/dal/detail/common.hpp"
 #include "oneapi/dal/detail/threading.hpp"
@@ -25,300 +26,230 @@
 
 namespace oneapi::dal::preview::triangle_counting::detail {
 
-template <typename Index>
-ONEDAL_EXPORT std::int64_t triangle_counting_global_scalar(const dal::detail::host_policy& policy,
-                                                           const Index* vertex_neighbors,
-                                                           const std::int64_t* edge_offsets,
-                                                           const Index* degrees,
-                                                           std::int64_t vertex_count,
-                                                           std::int64_t edge_count);
+template <typename Method, typename Task, typename Allocator, typename Topology>
+struct vertex_ranking_kernel_cpu {
+    inline vertex_ranking_result<Task> operator()(const dal::detail::host_policy& ctx,
+                                                  const detail::descriptor_base<Task>& desc,
+                                                  const Allocator& alloc,
+                                                  const Topology& t) const;
+};
 
-template <typename Index>
-ONEDAL_EXPORT std::int64_t triangle_counting_global_vector(const dal::detail::host_policy& policy,
-                                                           const Index* vertex_neighbors,
-                                                           const std::int64_t* edge_offsets,
-                                                           const Index* degrees,
-                                                           std::int64_t vertex_count,
-                                                           std::int64_t edge_count);
+struct scalar {};
+struct vector {};
+struct automatic {};
 
-template <typename Index>
-ONEDAL_EXPORT std::int64_t triangle_counting_global_vector_relabel(
-    const dal::detail::host_policy& policy,
-    const Index* vertex_neighbors,
-    const std::int64_t* edge_offsets,
-    const Index* degrees,
-    std::int64_t vertex_count,
-    std::int64_t edge_count);
+struct relabeled {};
 
-template <typename Index>
-ONEDAL_EXPORT array<std::int64_t> triangle_counting_local(
-    const dal::detail::host_policy& policy,
-    const dal::preview::detail::topology<Index>& data,
-    int64_t* triangles_local);
+template <typename Float, typename Task, typename Topology, typename... Param>
+struct triangle_counting {
+    vertex_ranking_result<Task> operator()(const dal::detail::host_policy& ctx,
+                                           const detail::descriptor_base<Task>& desc,
+                                           const Topology& t) const;
+};
+
+template <typename Float>
+struct triangle_counting<Float,
+                         task::local,
+                         dal::preview::detail::topology<std::int32_t>,
+                         automatic> {
+    array<std::int64_t> operator()(const dal::detail::host_policy& ctx,
+                                   const dal::preview::detail::topology<std::int32_t>& t,
+                                   std::int64_t* triangles_local) const;
+};
+
+template <typename Float>
+struct triangle_counting<Float,
+                         task::global,
+                         dal::preview::detail::topology<std::int32_t>,
+                         scalar> {
+    std::int64_t operator()(const dal::detail::host_policy& ctx,
+                            const dal::preview::detail::topology<std::int32_t>& t) const;
+};
+
+template <typename Float>
+struct triangle_counting<Float,
+                         task::global,
+                         dal::preview::detail::topology<std::int32_t>,
+                         vector> {
+    std::int64_t operator()(const dal::detail::host_policy& ctx,
+                            const dal::preview::detail::topology<std::int32_t>& t) const;
+};
+
+template <typename Float>
+struct triangle_counting<Float,
+                         task::global,
+                         dal::preview::detail::topology<std::int32_t>,
+                         vector,
+                         relabeled> {
+    std::int64_t operator()(const dal::detail::host_policy& ctx,
+                            const std::int32_t* vertex_neighbors,
+                            const std::int64_t* edge_offsets,
+                            const std::int32_t* degrees,
+                            std::int64_t vertex_count,
+                            std::int64_t edge_count) const;
+};
 
 ONEDAL_EXPORT std::int64_t compute_global_triangles(const dal::detail::host_policy& policy,
                                                     const array<std::int64_t>& local_triangles,
                                                     std::int64_t vertex_count);
 
-ONEDAL_EXPORT void sort_ids_by_degree(const dal::detail::host_policy& policy,
-                                      const std::int32_t* degrees,
-                                      std::pair<std::int32_t, std::size_t>* degree_id_pairs,
-                                      std::int64_t vertex_count);
+template <typename Allocator, typename Topology>
+struct triangle_counting_local {
+    inline array<std::int64_t> operator()(const dal::detail::host_policy& ctx,
+                                          const Allocator& alloc,
+                                          const Topology& t) {
+        const auto vertex_count = t.get_vertex_count();
 
-ONEDAL_EXPORT void fill_new_degrees_and_ids(
-    const dal::detail::host_policy& policy,
-    const std::pair<std::int32_t, std::size_t>* degree_id_pairs,
-    std::int32_t* new_ids,
-    std::int32_t* degrees_relabel,
-    std::int64_t vertex_count);
+        std::int64_t thread_cnt = dal::detail::threader_get_max_threads();
 
-ONEDAL_EXPORT void parallel_prefix_sum(const dal::detail::host_policy& policy,
-                                       const std::int32_t* degrees_relabel,
-                                       std::int64_t* offsets,
-                                       std::int64_t* part_prefix,
-                                       std::int64_t* local_sums,
-                                       std::int64_t block_size,
-                                       std::int64_t num_blocks,
-                                       std::int64_t vertex_count);
+        using int64_allocator_type =
+            typename std::allocator_traits<Allocator>::template rebind_alloc<std::int64_t>;
 
-ONEDAL_EXPORT void fill_relabeled_topology(const dal::detail::host_policy& policy,
-                                           const std::int32_t* vertex_neighbors,
-                                           const std::int64_t* edge_offsets,
-                                           std::int32_t* vertex_neighbors_relabel,
-                                           std::int64_t* edge_offsets_relabel,
-                                           std::int64_t* offsets,
-                                           const std::int32_t* new_ids,
-                                           std::int64_t vertex_count);
+        int64_allocator_type int64_allocator(alloc);
 
-template <typename Allocator>
-inline void relabel_by_greater_degree(const dal::detail::host_policy& ctx,
-                                      const std::int32_t* vertex_neighbors,
-                                      const std::int64_t* edge_offsets,
-                                      const std::int32_t* degrees,
-                                      std::int64_t vertex_count,
-                                      std::int64_t edge_count,
-                                      std::int32_t* vertex_neighbors_relabel,
-                                      std::int64_t* edge_offsets_relabel,
-                                      std::int32_t* degrees_relabel,
-                                      const Allocator& alloc) {
-    using int32_allocator_type =
-        typename std::allocator_traits<Allocator>::template rebind_alloc<std::int32_t>;
+        int64_t* triangles_local =
+            oneapi::dal::preview::detail::allocate(int64_allocator,
+                                                   (int64_t)thread_cnt * (int64_t)vertex_count);
 
-    using pair_allocator_type = typename std::allocator_traits<Allocator>::template rebind_alloc<
-        std::pair<std::int32_t, std::size_t>>;
+        auto arr_triangles =
+            triangle_counting<float, task::local, Topology, automatic>{}(ctx, t, triangles_local);
 
-    using int64_allocator_type =
-        typename std::allocator_traits<Allocator>::template rebind_alloc<std::int64_t>;
+        oneapi::dal::preview::detail::deallocate(int64_allocator,
+                                                 triangles_local,
+                                                 (int64_t)thread_cnt * (int64_t)vertex_count);
 
-    int64_allocator_type int64_allocator(alloc);
-    pair_allocator_type pair_allocator(alloc);
-    int32_allocator_type int32_allocator(alloc);
+        return arr_triangles;
+    }
+};
 
-    std::pair<std::int32_t, std::size_t>* degree_id_pairs =
+template <typename Allocator, typename Topology>
+struct vertex_ranking_kernel_cpu<method::ordered_count, task::global, Allocator, Topology> {
+    inline vertex_ranking_result<task::global> operator()(
+        const dal::detail::host_policy& ctx,
+        const detail::descriptor_base<task::global>& desc,
+        const Allocator& alloc,
+        const Topology& t) const {
+        const auto vertex_count = t.get_vertex_count();
+        if (vertex_count == 0) {
+            return vertex_ranking_result<task::global>();
+        }
+        const auto edge_count = t.get_edge_count();
+        const auto relabel = desc.get_relabel();
+        std::int64_t triangles = 0;
+        if (relabel == relabel::yes) {
+            const std::int32_t average_degree = edge_count / vertex_count;
+            const std::int32_t average_degree_sparsity_boundary = 4;
+            if (average_degree < average_degree_sparsity_boundary) {
+                triangles = triangle_counting<float, task::global, Topology, scalar>()(ctx, t);
+            }
+            else {
+                std::int32_t* g_vertex_neighbors_relabel = nullptr;
+                std::int64_t* g_edge_offsets_relabel = nullptr;
+                std::int32_t* g_degrees_relabel = nullptr;
 
-        oneapi::dal::preview::detail::allocate(pair_allocator, vertex_count);
+                using int32_allocator_type =
+                    typename std::allocator_traits<Allocator>::template rebind_alloc<std::int32_t>;
 
-    sort_ids_by_degree(ctx, degrees, degree_id_pairs, vertex_count);
+                using int64_allocator_type =
+                    typename std::allocator_traits<Allocator>::template rebind_alloc<std::int64_t>;
 
-    std::int32_t* new_ids = oneapi::dal::preview::detail::allocate(int32_allocator, vertex_count);
+                int64_allocator_type int64_allocator(alloc);
+                int32_allocator_type int32_allocator(alloc);
 
-    fill_new_degrees_and_ids(ctx, degree_id_pairs, new_ids, degrees_relabel, vertex_count);
+                g_vertex_neighbors_relabel =
+                    oneapi::dal::preview::detail::allocate(int32_allocator, edge_count * 2);
+                g_degrees_relabel =
+                    oneapi::dal::preview::detail::allocate(int32_allocator, vertex_count);
+                g_edge_offsets_relabel =
+                    oneapi::dal::preview::detail::allocate(int64_allocator, vertex_count + 1);
 
-    oneapi::dal::preview::detail::deallocate(pair_allocator, degree_id_pairs, vertex_count);
+                relabel_by_greater_degree<Allocator>{}(ctx,
+                                                       t,
+                                                       g_vertex_neighbors_relabel,
+                                                       g_edge_offsets_relabel,
+                                                       g_degrees_relabel,
+                                                       alloc);
 
-    std::int64_t* offsets =
-        oneapi::dal::preview::detail::allocate(int64_allocator, vertex_count + 1);
+                triangles = triangle_counting<float, task::global, Topology, vector, relabeled>()(
+                    ctx,
+                    g_vertex_neighbors_relabel,
+                    g_edge_offsets_relabel,
+                    g_degrees_relabel,
+                    vertex_count,
+                    edge_count);
 
-    const size_t block_size = 1 << 20;
-    const std::int64_t num_blocks = (vertex_count + block_size - 1) / block_size;
-    std::int64_t* local_sums = oneapi::dal::preview::detail::allocate(int64_allocator, num_blocks);
-    std::int64_t* part_prefix =
-        oneapi::dal::preview::detail::allocate(int64_allocator, num_blocks + 1);
+                oneapi::dal::preview::detail::deallocate(int32_allocator,
+                                                         g_vertex_neighbors_relabel,
+                                                         edge_count * 2);
+                oneapi::dal::preview::detail::deallocate(int32_allocator,
+                                                         g_degrees_relabel,
+                                                         vertex_count);
 
-    parallel_prefix_sum(ctx,
-                        degrees_relabel,
-                        offsets,
-                        part_prefix,
-                        local_sums,
-                        block_size,
-                        num_blocks,
-                        vertex_count);
-
-    oneapi::dal::preview::detail::deallocate(int64_allocator, local_sums, num_blocks);
-    oneapi::dal::preview::detail::deallocate(int64_allocator, part_prefix, num_blocks + 1);
-
-    fill_relabeled_topology(ctx,
-                            vertex_neighbors,
-                            edge_offsets,
-                            vertex_neighbors_relabel,
-                            edge_offsets_relabel,
-                            offsets,
-                            new_ids,
-                            vertex_count);
-
-    oneapi::dal::preview::detail::deallocate(int64_allocator, offsets, vertex_count + 1);
-    oneapi::dal::preview::detail::deallocate(int32_allocator, new_ids, vertex_count);
-}
-
-template <typename Allocator>
-inline vertex_ranking_result<task::global> triangle_counting_default_kernel(
-    const dal::detail::host_policy& ctx,
-    const detail::descriptor_base<task::global>& desc,
-    const Allocator& alloc,
-    const dal::preview::detail::topology<std::int32_t>& data) {
-    const auto g_edge_offsets = data._rows.get_data();
-    const auto g_vertex_neighbors = data._cols.get_data();
-    const auto g_degrees = data._degrees.get_data();
-    const auto g_vertex_count = data._vertex_count;
-    const auto g_edge_count = data._edge_count;
-
-    const auto relabel = desc.get_relabel();
-    std::int64_t triangles = 0;
-
-    if (relabel == relabel::yes) {
-        std::int32_t average_degree = g_edge_count / g_vertex_count;
-        const std::int32_t average_degree_sparsity_boundary = 4;
-        if (average_degree < average_degree_sparsity_boundary) {
-            triangles = triangle_counting_global_scalar(ctx,
-                                                        g_vertex_neighbors,
-                                                        g_edge_offsets,
-                                                        g_degrees,
-                                                        g_vertex_count,
-                                                        g_edge_count);
+                oneapi::dal::preview::detail::deallocate(int64_allocator,
+                                                         g_edge_offsets_relabel,
+                                                         vertex_count + 1);
+            }
         }
         else {
-            std::int32_t* g_vertex_neighbors_relabel = nullptr;
-            std::int64_t* g_edge_offsets_relabel = nullptr;
-            std::int32_t* g_degrees_relabel = nullptr;
-
-            using int32_allocator_type =
-                typename std::allocator_traits<Allocator>::template rebind_alloc<std::int32_t>;
-
-            using int64_allocator_type =
-                typename std::allocator_traits<Allocator>::template rebind_alloc<std::int64_t>;
-
-            int64_allocator_type int64_allocator(alloc);
-            int32_allocator_type int32_allocator(alloc);
-
-            g_vertex_neighbors_relabel =
-                oneapi::dal::preview::detail::allocate(int32_allocator,
-                                                       g_edge_offsets[g_vertex_count]);
-            g_degrees_relabel =
-                oneapi::dal::preview::detail::allocate(int32_allocator, g_vertex_count);
-            g_edge_offsets_relabel =
-                oneapi::dal::preview::detail::allocate(int64_allocator, g_vertex_count + 1);
-
-            relabel_by_greater_degree(ctx,
-                                      g_vertex_neighbors,
-                                      g_edge_offsets,
-                                      g_degrees,
-                                      g_vertex_count,
-                                      g_edge_count,
-                                      g_vertex_neighbors_relabel,
-                                      g_edge_offsets_relabel,
-                                      g_degrees_relabel,
-                                      alloc);
-
-            triangles = triangle_counting_global_vector_relabel(ctx,
-                                                                g_vertex_neighbors_relabel,
-                                                                g_edge_offsets_relabel,
-                                                                g_degrees_relabel,
-                                                                g_vertex_count,
-                                                                g_edge_count);
-
-            oneapi::dal::preview::detail::deallocate(int32_allocator,
-                                                     g_vertex_neighbors_relabel,
-                                                     g_edge_count);
-            oneapi::dal::preview::detail::deallocate(int32_allocator,
-                                                     g_degrees_relabel,
-                                                     g_vertex_count);
-
-            oneapi::dal::preview::detail::deallocate(int64_allocator,
-                                                     g_edge_offsets_relabel,
-                                                     g_vertex_count + 1);
+            const std::int32_t average_degree = edge_count / vertex_count;
+            const std::int32_t average_degree_sparsity_boundary = 4;
+            if (average_degree < average_degree_sparsity_boundary) {
+                triangles = triangle_counting<float, task::global, Topology, scalar>()(ctx, t);
+            }
+            else {
+                triangles = triangle_counting<float, task::global, Topology, vector>()(ctx, t);
+            }
         }
+
+        vertex_ranking_result<task::global> res;
+        res.set_global_rank(triangles);
+        return res;
     }
-    else {
-        std::int32_t average_degree = g_edge_count / g_vertex_count;
-        const std::int32_t average_degree_sparsity_boundary = 4;
-        if (average_degree < average_degree_sparsity_boundary) {
-            triangles = triangle_counting_global_scalar(ctx,
-                                                        g_vertex_neighbors,
-                                                        g_edge_offsets,
-                                                        g_degrees,
-                                                        g_vertex_count,
-                                                        g_edge_count);
+};
+
+template <typename Allocator, typename Topology>
+struct vertex_ranking_kernel_cpu<method::ordered_count, task::local, Allocator, Topology> {
+    inline vertex_ranking_result<task::local> operator()(
+        const dal::detail::host_policy& ctx,
+        const detail::descriptor_base<task::local>& desc,
+        const Allocator& alloc,
+        const Topology& t) const {
+        if (t.get_vertex_count() == 0) {
+            return vertex_ranking_result<task::local>();
         }
-        else {
-            triangles = triangle_counting_global_vector(ctx,
-                                                        g_vertex_neighbors,
-                                                        g_edge_offsets,
-                                                        g_degrees,
-                                                        g_vertex_count,
-                                                        g_edge_count);
-        }
+        auto local_triangles = triangle_counting_local<Allocator, Topology>()(ctx, alloc, t);
+
+        return vertex_ranking_result<task::local>().set_ranks(
+            dal::detail::homogen_table_builder{}
+                .reset(local_triangles, t.get_vertex_count(), 1)
+                .build());
     }
+};
 
-    vertex_ranking_result<task::global> res;
-    res.set_global_rank(triangles);
-    return res;
-}
+template <typename Allocator, typename Topology>
+struct vertex_ranking_kernel_cpu<method::ordered_count,
+                                 task::local_and_global,
+                                 Allocator,
+                                 Topology> {
+    inline vertex_ranking_result<task::local_and_global> operator()(
+        const dal::detail::host_policy& ctx,
+        const detail::descriptor_base<task::local_and_global>& desc,
+        const Allocator& alloc,
+        const Topology& t) const {
+        const auto vertex_count = t.get_vertex_count();
+        if (vertex_count == 0) {
+            return vertex_ranking_result<task::local_and_global>();
+        }
+        auto local_triangles = triangle_counting_local<Allocator, Topology>()(ctx, alloc, t);
 
-template <typename Allocator>
-inline array<std::int64_t> triangle_counting_local_default_kernel(
-    const dal::detail::host_policy& ctx,
-    const Allocator& alloc,
-    const dal::preview::detail::topology<std::int32_t>& data) {
-    const auto g_vertex_count = data._vertex_count;
+        std::int64_t total_s = compute_global_triangles(ctx, local_triangles, vertex_count);
 
-    int thread_cnt = dal::detail::threader_get_max_threads();
-
-    using int64_allocator_type =
-        typename std::allocator_traits<Allocator>::template rebind_alloc<std::int64_t>;
-
-    int64_allocator_type int64_allocator(alloc);
-
-    int64_t* triangles_local =
-        oneapi::dal::preview::detail::allocate(int64_allocator,
-                                               (int64_t)thread_cnt * (int64_t)g_vertex_count);
-
-    auto arr_triangles = triangle_counting_local(ctx, data, triangles_local);
-
-    oneapi::dal::preview::detail::deallocate(int64_allocator,
-                                             triangles_local,
-                                             (int64_t)thread_cnt * (int64_t)g_vertex_count);
-
-    return arr_triangles;
-}
-
-template <typename Allocator>
-inline vertex_ranking_result<task::local> triangle_counting_default_kernel(
-    const dal::detail::host_policy& ctx,
-    const detail::descriptor_base<task::local>& desc,
-    const Allocator& alloc,
-    const dal::preview::detail::topology<std::int32_t>& data) {
-    auto local_triangles = triangle_counting_local_default_kernel(ctx, alloc, data);
-
-    return vertex_ranking_result<task::local>().set_ranks(
-        dal::detail::homogen_table_builder{}.reset(local_triangles, data._vertex_count, 1).build());
-}
-
-template <typename Allocator>
-inline vertex_ranking_result<task::local_and_global> triangle_counting_default_kernel(
-    const dal::detail::host_policy& ctx,
-    const detail::descriptor_base<task::local_and_global>& desc,
-    const Allocator& alloc,
-    const dal::preview::detail::topology<std::int32_t>& data) {
-    const auto vertex_count = data._vertex_count;
-
-    auto local_triangles = triangle_counting_local_default_kernel(ctx, alloc, data);
-
-    std::int64_t total_s = compute_global_triangles(ctx, local_triangles, vertex_count);
-
-    return vertex_ranking_result<task::local_and_global>()
-        .set_ranks(
-            dal::detail::homogen_table_builder{}.reset(local_triangles, vertex_count, 1).build())
-        .set_global_rank(total_s);
-}
+        return vertex_ranking_result<task::local_and_global>()
+            .set_ranks(dal::detail::homogen_table_builder{}
+                           .reset(local_triangles, vertex_count, 1)
+                           .build())
+            .set_global_rank(total_s);
+    }
+};
 
 } // namespace oneapi::dal::preview::triangle_counting::detail
