+++ cpp/oneapi/dal/backend/primitives/BUILD
@@ -24,6 +24,7 @@ dal_collect_modules(
         "reduction",
         "selection",
         "sort",
+        "intersection",
         "rng",
         "distance"
     ],
