+++ cpp/oneapi/dal/backend/primitives/intersection/BUILD
@@ -1,0 +1,13 @@
+package(default_visibility = ["//visibility:public"])
+load("@onedal//dev/bazel:dal.bzl",
+    "dal_module",
+    "dal_test_suite",
+)
+
+dal_module(
+    name = "intersection",
+    auto = True,
+    dal_deps = [
+    ],
+)
+
