+++ cpp/oneapi/dal/graph/service_functions.hpp
@@ -29,71 +29,71 @@ namespace oneapi::dal::preview {
 /// Returns the number of vertices in the graph
 ///
 /// @tparam Graph  Type of the graph
-/// @param [in]   graph  Input graph object
+/// @param [in]   g  Input graph object
 /// @return The number of vertices in the graph
 template <typename Graph>
-constexpr auto get_vertex_count(const Graph &graph) noexcept -> vertex_size_type<Graph>;
+constexpr auto get_vertex_count(const Graph &g) noexcept -> vertex_size_type<Graph>;
 
 /// Returns the number of edges in the graph
 ///
 /// @tparam Graph  Type of the graph
-/// @param [in]   graph  Input graph object
+/// @param [in]   g  Input graph object
 ///
 /// @return The number of edges in the graph
 template <typename Graph>
-constexpr auto get_edge_count(const Graph &graph) noexcept -> edge_size_type<Graph>;
+constexpr auto get_edge_count(const Graph &g) noexcept -> edge_size_type<Graph>;
 
 /// Returns the degree for the specified vertex
 ///
 /// @tparam Graph  Type of the graph
-/// @param [in]   graph  Input graph object
-/// @param [in]   vertex Identifier of the vertex
+/// @param [in]   g  Input graph object
+/// @param [in]   u Identifier of the vertex
 ///
-/// @return The degree of the vertex
+/// @return The degree of the vertex u
 template <typename Graph>
-constexpr auto get_vertex_degree(const Graph &graph, vertex_type<Graph> vertex)
+constexpr auto get_vertex_degree(const Graph &g, vertex_type<Graph> u)
     -> vertex_edge_size_type<Graph>;
 
 /// Returns the range of the vertex neighbors for the specified vertex
 ///
 /// @tparam Graph  Type of the graph
-/// @param [in]   graph  Input graph object
-/// @param [in]   vertex Identifier of the vertex
+/// @param [in]   g  Input graph object
+/// @param [in]   u Identifier of the vertex
 ///
-/// @return The range of the vertex neighbors
+/// @return The range of the vertex u neighbors
 template <typename Graph>
-constexpr auto get_vertex_neighbors(const Graph &graph, vertex_type<Graph> vertex)
+constexpr auto get_vertex_neighbors(const Graph &g, vertex_type<Graph> u)
     -> const_vertex_edge_range_type<Graph>;
 
 //Functions implementation
 template <typename Graph>
-constexpr auto get_vertex_count(const Graph &graph) noexcept -> vertex_size_type<Graph> {
-    return detail::get_vertex_count_impl(graph);
+constexpr auto get_vertex_count(const Graph &g) noexcept -> vertex_size_type<Graph> {
+    return detail::get_vertex_count_impl(g);
 }
 
 template <typename Graph>
-constexpr auto get_edge_count(const Graph &graph) noexcept -> edge_size_type<Graph> {
-    return detail::get_edge_count_impl(graph);
+constexpr auto get_edge_count(const Graph &g) noexcept -> edge_size_type<Graph> {
+    return detail::get_edge_count_impl(g);
 }
 
 template <typename Graph>
-constexpr auto get_vertex_degree(const Graph &graph, vertex_type<Graph> vertex)
+constexpr auto get_vertex_degree(const Graph &g, vertex_type<Graph> u)
     -> vertex_edge_size_type<Graph> {
-    if (vertex < 0 || (vertex_size_type<Graph>)vertex >= detail::get_vertex_count_impl(graph)) {
+    if (u < 0 || (vertex_size_type<Graph>)u >= detail::get_vertex_count_impl(g)) {
         throw out_of_range(dal::detail::error_messages::
                                vertex_index_out_of_range_expect_from_zero_to_vertex_count());
     }
-    return detail::get_vertex_degree_impl(graph, vertex);
+    return detail::get_vertex_degree_impl(g, u);
 }
 
 template <typename Graph>
-constexpr auto get_vertex_neighbors(const Graph &graph, vertex_type<Graph> vertex)
+constexpr auto get_vertex_neighbors(const Graph &g, vertex_type<Graph> u)
     -> const_vertex_edge_range_type<Graph> {
-    if (vertex < 0 || (vertex_size_type<Graph>)vertex >= detail::get_vertex_count_impl(graph)) {
+    if (u < 0 || (vertex_size_type<Graph>)u >= detail::get_vertex_count_impl(g)) {
         throw out_of_range(dal::detail::error_messages::
                                vertex_index_out_of_range_expect_from_zero_to_vertex_count());
     }
-    return detail::get_vertex_neighbors_impl(graph, vertex);
+    return detail::get_vertex_neighbors_impl(g, u);
 }
 
 } // namespace oneapi::dal::preview
